Move worlds in the separate folder
