+++ tests/test_actions.py
@@ -5,7 +5,7 @@ Author: Maria Rigaki - maria.rigaki@fel.cvut.cz
 import sys
 from os import path
 sys.path.append( path.dirname(path.dirname( path.abspath(__file__) ) ))
-from env.network_security_game import NetworkSecurityEnvironment
+from env.worlds.network_security_game import NetworkSecurityEnvironment
 import env.game_components as components
 import pytest
